#17442 - Generate idLote without using concurrency control

+++ NFe.Classes/Servicos/Autorizacao/enviNFe3.cs
@@ -39,7 +39,7 @@ namespace NFe.Classes.Servicos.Autorizacao
     [XmlRoot(ElementName = "enviNFe", Namespace = "http://www.portalfiscal.inf.br/nfe")]
     public class enviNFe3
     {
-        public enviNFe3(string versao, int idLote, IndicadorSincronizacao indSinc, List<NFe> nFe)
+        public enviNFe3(string versao, long idLote, IndicadorSincronizacao indSinc, List<NFe> nFe)
         {
             this.versao = versao;
             this.idLote = idLote;
@@ -60,7 +60,7 @@ namespace NFe.Classes.Servicos.Autorizacao
         /// <summary>
         ///     AP03 - Identificador de controle do envio do lote.
         /// </summary>
-        public int idLote { get; set; }
+        public long idLote { get; set; }
 
         /// <summary>
         ///     AP03a - Indicador de Sincronização
